# packages/server/openapi.yaml
openapi: 3.0.0
info:
  title: Worlds API
  version: 1.0.0
  description: API for managing and querying RDF-based "worlds".
servers:
  - url: /v1
    description: V1 API
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    TripleSearchResult:
      type: object
      properties:
        subject:
          type: string
        predicate:
          type: string
        object:
          type: string
        vecRank:
          type: number
          nullable: true
        ftsRank:
          type: number
          nullable: true
        score:
          type: number
        worldId:
          type: string
        organizationId:
          type: string
      required:
        - subject
        - predicate
        - object
        - score
    SparqlResults:
      type: object
      oneOf:
        - $ref: "#/components/schemas/SparqlSelectResults"
        - $ref: "#/components/schemas/SparqlAskResults"
        - $ref: "#/components/schemas/SparqlQuadsResults"
    SparqlSelectResults:
      type: object
      properties:
        head:
          type: object
          properties:
            vars:
              type: array
              items:
                type: string
            link:
              type: array
              items:
                type: string
              nullable: true
        results:
          type: object
          properties:
            bindings:
              type: array
              items:
                type: object
                additionalProperties:
                  $ref: "#/components/schemas/SparqlValue"
    SparqlAskResults:
      type: object
      properties:
        head:
          type: object
          properties:
            link:
              type: array
              items:
                type: string
              nullable: true
        boolean:
          type: boolean
    SparqlQuadsResults:
      type: object
      properties:
        head:
          type: object
          properties:
            link:
              type: array
              items:
                type: string
              nullable: true
        results:
          type: object
          properties:
            quads:
              type: array
              items:
                $ref: "#/components/schemas/SparqlQuad"
    SparqlValue:
      type: object
      discriminator:
        propertyName: type
      properties:
        type:
          type: string
          enum: [uri, bnode, literal]
        value:
          type: string
      required:
        - type
        - value
      oneOf:
        - properties:
            type:
              const: uri
        - properties:
            type:
              const: bnode
        - properties:
            type:
              const: literal
            xml:lang:
              type: string
            datatype:
              type: string
    SparqlQuad:
      type: object
      properties:
        subject:
          type: object
          properties:
            type:
              type: string
              enum: [uri, bnode]
            value:
              type: string
        predicate:
          type: object
          properties:
            type:
              type: string
              const: uri
            value:
              type: string
        object:
          $ref: "#/components/schemas/SparqlValue"
        graph:
          type: object
          properties:
            type:
              type: string
              enum: [default, uri]
            value:
              type: string
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
      required:
        - error
security:
  - bearerAuth: []
paths:
...
  /worlds/{world}/search:
    get:
      summary: Semantic search
      parameters:
        - name: world
          in: path
          required: true
          schema:
            type: string
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: subjects
          in: query
          schema:
            type: array
            items:
              type: string
        - name: predicates
          in: query
          schema:
            type: array
            items:
              type: string
        - name: limit
          in: query
          schema:
            type: integer
            maximum: 100
            default: 20
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TripleSearchResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /worlds/{world}/sparql:
    get:
      summary: Execute SPARQL query or get service description
      parameters:
        - name: world
          in: path
          required: true
          schema:
            type: string
        - name: query
          in: query
          schema:
            type: string
        - name: default-graph-uri
          in: query
          schema:
            type: array
            items:
              type: string
        - name: named-graph-uri
          in: query
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: Success
          content:
            application/sparql-results+json:
              schema:
                $ref: "#/components/schemas/SparqlResults"
            text/turtle:
              description: SPARQL Service Description
              schema: { type: string }
            application/rdf+xml:
              description: SPARQL Service Description
              schema: { type: string }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      summary: Execute SPARQL query or update
      parameters:
        - name: world
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                query:
                  type: string
                default-graph-uri:
                  type: array
                  items:
                    type: string
                named-graph-uri:
                  type: array
                  items:
                    type: string
          application/sparql-query:
            schema: { type: string }
          application/sparql-update:
            schema: { type: string }
      responses:
        "200":
          description: Success
          content:
            application/sparql-results+json:
              schema:
                $ref: "#/components/schemas/SparqlResults"
        "204":
          description: Update success
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"